requirements.txt
streamlit
google-generativeai
python-dotenv
langchain
PyPDF2
langchain
chromadb
faiss-cpu
langchain_google_genaipip